requirements.txt
pandas
scikit-learn
matplotlib
jupyter
umap-learn
hdbscan